Changes from PR
1) File renamed by mistake and correct them
